Split three oversized components into focused units (#226, #229, #233)
GlobalTab (952→371 lines): edit form state, validation, and JSX extracted
to ContactEditForm.tsx. SOCIAL_TYPES/SOCIAL_META/KNOWN_LINK_TYPES constants
now live in ContactEditForm and are exported for GlobalTab's view mode.

SettingsView (1043→581 lines): Profile+Security extracted to ProfileSection,
Backup+AutoBackup extracted to BackupSection. Toggle component extracted to
SettingsToggle.tsx so section files can share it.

ProjectView (977→714 lines): bulk-selection bar extracted to BulkBar (owns
its own confirmDelete/confirmRemove/bulkWorking state, resets on checkedCount
change). Filter+sort logic extracted to hooks/useProjectContacts. SortState
type added to useSortFilter exports.
