+++ src/renderer/src/contacts/GlobalTab.tsx
@@ -1,23 +1,15 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import type { ContactDetail as ContactDetailType, ContactAlertRss, Project, User } from '@shared/types';
 import Button from '../shell/Button';
 import GlobalLogSection from './GlobalLogSection';
 import GlobalRemindersSection from './GlobalRemindersSection';
-import DynamicList, { useDragReorder } from './DynamicList';
-import {
-  isGoogleAlertUrl,
-  hasDisallowedPhoneChars,
-  normalizePhoneForComparison,
-  sanitizeOtherLabel,
-  OTHER_LABEL_MAX,
-} from './contactValidation';
-import { NON_OTHER_SOCIAL_TYPES, SOCIAL_TYPES, SOCIAL_META, localToday } from './contactConstants';
-import type { SocialType, NonOtherSocialType } from './contactConstants';
-import { useContactFieldValidation } from './useContactFieldValidation';
-import { CalendarPicker } from '../views/CalendarPicker';
+import ContactEditForm, { SOCIAL_TYPES, SOCIAL_META, KNOWN_LINK_TYPES } from './ContactEditForm';
+import { isGoogleAlertUrl } from './contactValidation';
 import RssAlertPanel from './RssAlertPanel';
 import ScreenshotPanel from './ScreenshotPanel';
 import { linkifyText } from '../utils/linkify';
+import { HANDLE_TYPES, HANDLE_META } from './handleMeta';
+import type { HandleType } from './handleMeta';
 import './AddContactModal.css';
 import './ContactDetail.css';
 
@@ -31,12 +23,6 @@ interface Props {
   user?: User | null;
 }
 
-import { HANDLE_TYPES, HANDLE_META } from './handleMeta';
-import type { HandleType } from './handleMeta';
-
-const KNOWN_LINK_TYPES = new Set<string>([...SOCIAL_TYPES, 'website']);
-
-
 export default function GlobalTab({ contact, allProjects, onRefresh, onMembershipChanged, onDeleted, onEditingChange, user }: Props) {
   const [editing, setEditing] = useState(false);
 
@@ -44,15 +30,13 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
     setEditing(value);
     onEditingChange?.(value);
   }
-  const [saving, setSaving] = useState(false);
+
   const [confirmDelete, setConfirmDelete] = useState(false);
   const [addingToProject, setAddingToProject] = useState('');
   const [confirmRemoveProjectId, setConfirmRemoveProjectId] = useState<string | null>(null);
   const [alertRssList, setAlertRssList] = useState<ContactAlertRss[]>([]);
   const [waybackStatus, setWaybackStatus] = useState<Map<string, 'pending' | 'failed'>>(new Map());
-  const formRef = useRef<HTMLDivElement>(null);
-  const mounted = useRef(true);
-  useEffect(() => { mounted.current = true; return () => { mounted.current = false; }; }, []);
+  const [newRssUrl, setNewRssUrl] = useState('');
 
   useEffect(() => {
     return window.sourcerer.onWaybackStatus(({ contactId, url, status }) => {
@@ -69,7 +53,6 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
     });
   }, [contact.id]);
 
-  // Clear pending status when contact reloads with a wayback_url
   useEffect(() => {
     const archivedUrls = new Set(contact.links.filter((l) => l.wayback_url).map((l) => l.url));
     if (archivedUrls.size === 0) return;
@@ -81,58 +64,6 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
     });
   }, [contact.links]);
 
-  // Edit form state
-  const [editName, setEditName] = useState('');
-  const [editOrg, setEditOrg] = useState('');
-  const [editTitle, setEditTitle] = useState('');
-  const [editDob, setEditDob] = useState('');
-  const [editNotes, setEditNotes] = useState('');
-  const [editHandles, setEditHandles] = useState<Array<{ type: HandleType; handle: string }>>([]);
-  const [editEmails, setEditEmails] = useState<Array<{ email: string; label: string }>>([]);
-  const [editPhones, setEditPhones] = useState<Array<{ phone: string; label: string }>>([]);
-  const [editSocials, setEditSocials] = useState<Record<NonOtherSocialType, string[]>>({
-    linkedin: [], x: [], instagram: [], facebook: [],
-  });
-  const [editOtherSocials, setEditOtherSocials] = useState<Array<{ url: string; label: string }>>([]);
-  const { getDragProps: emailDragProps, handleProps: emailHandleProps } = useDragReorder(editEmails, setEditEmails);
-  const { getDragProps: phoneDragProps, handleProps: phoneHandleProps } = useDragReorder(editPhones, setEditPhones);
-  const { getDragProps: otherSocialDragProps, handleProps: otherSocialHandleProps } = useDragReorder(editOtherSocials, setEditOtherSocials);
-  const [editWebsites, setEditWebsites] = useState<string[]>([]);
-  const [newRssUrl, setNewRssUrl] = useState('');
-  const urlValues = useMemo(
-    () => [
-      ...editWebsites,
-      ...NON_OTHER_SOCIAL_TYPES.flatMap((t) => editSocials[t]),
-      ...editOtherSocials.map((e) => e.url),
-    ],
-    [editWebsites, editSocials, editOtherSocials],
-  );
-
-  const {
-    emailCollisions,
-    phoneCollisions,
-    emailFormatWarnings,
-    phoneFormatWarnings,
-    urlFormatWarnings,
-    emailDuplicates,
-    phoneDuplicates,
-    urlDuplicates,
-    checkEmailBlur,
-    checkPhoneBlur,
-    clearEmailWarnings,
-    clearPhoneWarnings,
-    updatePhoneFormatWarning,
-    clearUrlWarning,
-    updateUrlFormatWarning,
-    resetAll,
-  } = useContactFieldValidation({
-    excludeId: contact.id,
-    mounted,
-    emails: editEmails,
-    phones: editPhones,
-    urlValues,
-  });
-
   useEffect(() => {
     let cancelled = false;
     window.sourcerer.listAlertRss(contact.id).then((list) => {
@@ -141,78 +72,11 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
     return () => { cancelled = true; };
   }, [contact.id]);
 
-
-  function setSocial(type: NonOtherSocialType, values: string[]) {
-    setEditSocials((prev) => ({ ...prev, [type]: values }));
-  }
-
-  function startEdit() {
-    setEditName(contact.name);
-    setEditOrg(contact.organization ?? '');
-    setEditTitle(contact.title ?? '');
-    setEditDob(contact.dob ?? '');
-    setEditNotes(contact.notes ?? '');
-    setEditHandles(contact.handles.map((h) => ({ type: h.type, handle: h.handle })));
-    setEditEmails(contact.emails.map((e) => ({ email: e.email, label: e.label ?? '' })));
-    setEditPhones(contact.phones.map((p) => ({ phone: p.phone, label: p.label ?? '' })));
-    const socialsByType = Object.fromEntries(
-      NON_OTHER_SOCIAL_TYPES.map((type) => [
-        type,
-        contact.links.filter((l) => l.type === type).map((l) => l.url),
-      ]),
-    ) as Record<NonOtherSocialType, string[]>;
-    setEditSocials(socialsByType);
-    setEditOtherSocials(
-      contact.links.filter((l) => l.type === 'other').map((l) => ({ url: l.url, label: l.label ?? '' })),
-    );
-    setEditWebsites(contact.links.filter((l) => l.type === 'website').map((l) => l.url));
+  // Reset edit mode when contact changes
+  useEffect(() => {
+    setEditingAndNotify(false);
     setNewRssUrl('');
-    resetAll();
-    setEditingAndNotify(true);
-  }
-
-  async function handleSave() {
-    if (!editName.trim()) return;
-    if (emailDuplicates.size > 0 || phoneDuplicates.size > 0 || urlDuplicates.size > 0) {
-      formRef.current?.querySelector<HTMLElement>('.ac-collision-warn')?.scrollIntoView({ behavior: 'smooth', block: 'center' });
-      return;
-    }
-    setSaving(true);
-    try {
-      const links = [
-        ...NON_OTHER_SOCIAL_TYPES.flatMap((type) =>
-          editSocials[type].filter((u) => u.trim()).map((url) => ({ type, url })),
-        ),
-        ...editOtherSocials
-          .filter((e) => e.url.trim())
-          .map((e) => {
-            const label = sanitizeOtherLabel(e.label);
-            return { type: 'other' as const, url: e.url, ...(label ? { label } : {}) };
-          }),
-        ...editWebsites.filter((u) => u.trim()).map((url) => ({ type: 'website' as const, url })),
-      ];
-      await window.sourcerer.updateContact({
-        id: contact.id,
-        name: editName,
-        organization: editOrg,
-        title: editTitle,
-        dob: editDob || undefined,
-        notes: editNotes,
-        emails: editEmails.filter((e) => e.email.trim()).map((e) => ({ email: e.email, label: e.label.trim() || undefined })),
-        phones: editPhones.filter((p) => p.phone?.trim()).map((p) => ({ phone: p.phone, label: p.label.trim() || undefined })),
-        links,
-        handles: editHandles.filter((h) => h.handle.trim() && h.type.trim()),
-      });
-      const pendingRss = newRssUrl.trim();
-      if (pendingRss && isGoogleAlertUrl(pendingRss) && !alertRssList.some((f) => f.rss_url === pendingRss)) {
-        await window.sourcerer.addAlertRss(contact.id, pendingRss);
-      }
-      onRefresh();
-      setEditingAndNotify(false);
-    } finally {
-      setSaving(false);
-    }
-  }
+  }, [contact.id]);
 
   async function handleAddToProject() {
     if (!addingToProject) return;
@@ -257,11 +121,9 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
     } catch { /* item stays in list */ }
   }
 
-
-  // View mode: group links by type
   const socialLinks = Object.fromEntries(
     SOCIAL_TYPES.map((type) => [type, contact.links.filter((l) => l.type === type)]),
-  ) as Record<SocialType, typeof contact.links>;
+  ) as Record<typeof SOCIAL_TYPES[number], typeof contact.links>;
   const websiteLinks = contact.links.filter((l) => l.type === 'website');
   const otherLinks = contact.links.filter((l) => !KNOWN_LINK_TYPES.has(l.type));
   const contactProjectIds = new Set(contact.projects.map((p) => p.id));
@@ -269,357 +131,17 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
 
   if (editing) {
     return (
-      <div className="detail-body" ref={formRef}>
-        <div className="form-field">
-          <label className="form-label">Name <span className="form-required">*</span></label>
-          <input className="form-input" value={editName} onChange={(e) => setEditName(e.target.value)} autoFocus />
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Organization</label>
-          <input className="form-input" value={editOrg} onChange={(e) => setEditOrg(e.target.value)} />
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Title / Role</label>
-          <input className="form-input" value={editTitle} onChange={(e) => setEditTitle(e.target.value)} placeholder="e.g. Senior Editor" />
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Date of birth</label>
-          <CalendarPicker
-            label="Select date"
-            value={editDob}
-            onChange={setEditDob}
-            showYear
-            maxDate={localToday()}
-          />
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Email</label>
-          {editEmails.map((entry, i) => (
-            <div
-              key={i}
-              {...emailDragProps(i)}
-            >
-              <div className="ac-phone-row">
-                <span className="ac-drag-handle" {...emailHandleProps}>⠿</span>
-                <input
-                  className="form-input"
-                  value={entry.email}
-                  placeholder="email@example.com"
-                  disabled={saving}
-                  onChange={(e) => {
-                    const prev = entry.email.trim();
-                    const next = [...editEmails];
-                    next[i] = { ...next[i], email: e.target.value };
-                    setEditEmails(next);
-                    if (prev) clearEmailWarnings(prev);
-                  }}
-                  onBlur={() => checkEmailBlur(entry.email.trim())}
-                />
-                <input
-                  className="form-input"
-                  value={entry.label}
-                  placeholder="label"
-                  disabled={saving}
-                  onChange={(e) => {
-                    const next = [...editEmails];
-                    next[i] = { ...next[i], label: e.target.value };
-                    setEditEmails(next);
-                  }}
-                />
-                <button
-                  className="ac-remove"
-                  type="button"
-                  onClick={() => setEditEmails(editEmails.filter((_, j) => j !== i))}
-                ></button>
-              </div>
-              {entry.email.trim() && emailFormatWarnings[entry.email.trim()] && (
-                <div className="ac-collision-warn">
-                  ⚠ Invalid email address
-                </div>
-              )}
-              {entry.email.trim() && !emailFormatWarnings[entry.email.trim()] && emailDuplicates.has(entry.email.trim().toLowerCase()) && (
-                <div className="ac-collision-warn">⚠ Already added</div>
-              )}
-              {entry.email.trim() && !emailFormatWarnings[entry.email.trim()] && !emailDuplicates.has(entry.email.trim().toLowerCase()) && emailCollisions[entry.email.trim()] && (
-                <div className="ac-collision-warn">
-                  Already on: <span>{emailCollisions[entry.email.trim()]}</span>
-                </div>
-              )}
-            </div>
-          ))}
-          <div>
-            <Button
-              variant="ghost"
-              type="button"
-              onClick={() => setEditEmails([...editEmails, { email: '', label: '' }])}
-            >
-              + Add
-            </Button>
-          </div>
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Phone</label>
-          {editPhones.map((entry, i) => (
-            <div
-              key={i}
-              {...phoneDragProps(i)}
-            >
-              <div className="ac-phone-row">
-                <span className="ac-drag-handle" {...phoneHandleProps}>⠿</span>
-                <input
-                  className="form-input"
-                  value={entry.phone}
-                  placeholder="+1 555 000 0000"
-                  disabled={saving}
-                  onChange={(e) => {
-                    const prev = entry.phone.trim();
-                    const next = [...editPhones];
-                    next[i] = { ...next[i], phone: e.target.value };
-                    setEditPhones(next);
-                    const newVal = e.target.value.trim();
-                    if (prev && prev !== newVal) clearPhoneWarnings(prev);
-                    if (newVal) updatePhoneFormatWarning(newVal);
-                  }}
-                  onBlur={() => checkPhoneBlur(entry.phone.trim())}
-                />
-                <input
-                  className="form-input"
-                  value={entry.label}
-                  placeholder="label"
-                  disabled={saving}
-                  onChange={(e) => {
-                    const next = [...editPhones];
-                    next[i] = { ...next[i], label: e.target.value };
-                    setEditPhones(next);
-                  }}
-                />
-                <button
-                  className="ac-remove"
-                  type="button"
-                  onClick={() => setEditPhones(editPhones.filter((_, j) => j !== i))}
-                ></button>
-              </div>
-              {entry.phone.trim() && phoneFormatWarnings[entry.phone.trim()] && (
-                <div className="ac-collision-warn">
-                  {hasDisallowedPhoneChars(entry.phone.trim())
-                    ? '⚠ Phone contains invalid characters'
-                    : '⚠ Invalid phone number'}
-                </div>
-              )}
-              {entry.phone.trim() && !phoneFormatWarnings[entry.phone.trim()] && phoneDuplicates.has(normalizePhoneForComparison(entry.phone)) && (
-                <div className="ac-collision-warn">⚠ Already added</div>
-              )}
-              {entry.phone.trim() && !phoneFormatWarnings[entry.phone.trim()] && !phoneDuplicates.has(normalizePhoneForComparison(entry.phone)) && phoneCollisions[entry.phone.trim()] && (
-                <div className="ac-collision-warn">
-                  Already on: <span>{phoneCollisions[entry.phone.trim()]}</span>
-                </div>
-              )}
-            </div>
-          ))}
-          <div>
-            <Button
-              variant="ghost"
-              type="button"
-              onClick={() => setEditPhones([...editPhones, { phone: '', label: '' }])}
-            >
-              + Add
-            </Button>
-          </div>
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Messaging</label>
-          {editHandles.map((entry, i) => (
-            <div key={i} className="ac-phone-row">
-              <select
-                className="form-input ac-handle-type"
-                value={HANDLE_TYPES.includes(entry.type as HandleType) ? entry.type : 'other'}
-                onChange={(e) => {
-                  const next = [...editHandles];
-                  next[i] = { ...next[i], type: e.target.value as HandleType };
-                  setEditHandles(next);
-                }}
-              >
-                {HANDLE_TYPES.map((t) => (
-                  <option key={t} value={t}>{HANDLE_META[t].label}</option>
-                ))}
-              </select>
-              <input
-                className="form-input"
-                value={entry.handle}
-                placeholder={HANDLE_META[(HANDLE_TYPES.includes(entry.type as HandleType) ? entry.type : 'other') as HandleType].placeholder}
-                onChange={(e) => {
-                  const next = [...editHandles];
-                  next[i] = { ...next[i], handle: e.target.value };
-                  setEditHandles(next);
-                }}
-              />
-              <button
-                className="ac-remove"
-                type="button"
-                onClick={() => setEditHandles(editHandles.filter((_, j) => j !== i))}
-              ></button>
-            </div>
-          ))}
-          <div>
-            <Button
-              variant="ghost"
-              type="button"
-              onClick={() => setEditHandles([...editHandles, { type: 'signal', handle: '' }])}
-            >
-              + Add
-            </Button>
-          </div>
-        </div>
-
-        {NON_OTHER_SOCIAL_TYPES.map((type) => (
-          <div key={type} className="form-field">
-            <label className="form-label">{SOCIAL_META[type].label}</label>
-            <DynamicList
-              enableDragReorder
-              values={editSocials[type]}
-              placeholder={SOCIAL_META[type].placeholder}
-              onChange={(vals) => setSocial(type, vals)}
-              onChangeItem={(oldVal) => { if (oldVal) clearUrlWarning(oldVal); }}
-              onBlurItem={(val) => { if (val) updateUrlFormatWarning(val); }}
-              warnings={Object.fromEntries(
-                editSocials[type]
-                  .map((v) => v.trim())
-                  .filter(Boolean)
-                  .flatMap((v) => {
-                    if (urlFormatWarnings[v]) return [[v, '⚠ Invalid URL']];
-                    if (urlDuplicates.has(v)) return [[v, '⚠ Already added']];
-                    return [];
-                  })
-              )}
-            />
-          </div>
-        ))}
-
-        <div className="form-field">
-          <label className="form-label">Other social</label>
-          {editOtherSocials.map((entry, i) => (
-            <div
-              key={i}
-              {...otherSocialDragProps(i)}
-            >
-              <div className="ac-other-social-row">
-                <span className="ac-drag-handle" {...otherSocialHandleProps}>⠿</span>
-                <input
-                  className="form-input"
-                  value={entry.label}
-                  placeholder="Platform (e.g. TikTok)"
-                  maxLength={OTHER_LABEL_MAX}
-                  onChange={(e) => {
-                    const next = [...editOtherSocials];
-                    next[i] = { ...next[i], label: e.target.value };
-                    setEditOtherSocials(next);
-                  }}
-                  onBlur={(e) => {
-                    const next = [...editOtherSocials];
-                    next[i] = { ...next[i], label: sanitizeOtherLabel(e.target.value) };
-                    setEditOtherSocials(next);
-                  }}
-                />
-                <input
-                  className="form-input"
-                  value={entry.url}
-                  placeholder="https://…"
-                  onChange={(e) => {
-                    const prev = entry.url.trim();
-                    const next = [...editOtherSocials];
-                    next[i] = { ...next[i], url: e.target.value };
-                    setEditOtherSocials(next);
-                    if (prev) clearUrlWarning(prev);
-                  }}
-                  onBlur={() => {
-                    const val = entry.url.trim();
-                    if (val) updateUrlFormatWarning(val);
-                  }}
-                />
-                <button
-                  className="ac-remove"
-                  type="button"
-                  onClick={() => setEditOtherSocials(editOtherSocials.filter((_, j) => j !== i))}
-                ></button>
-              </div>
-              {entry.url.trim() && urlFormatWarnings[entry.url.trim()] && (
-                <div className="ac-collision-warn">⚠ Invalid URL</div>
-              )}
-              {entry.url.trim() && !urlFormatWarnings[entry.url.trim()] && urlDuplicates.has(entry.url.trim()) && (
-                <div className="ac-collision-warn">⚠ Already added</div>
-              )}
-            </div>
-          ))}
-          <div>
-            <Button
-              variant="ghost"
-              type="button"
-              onClick={() => setEditOtherSocials([...editOtherSocials, { url: '', label: '' }])}
-            >
-              + Add
-            </Button>
-          </div>
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Website</label>
-          <DynamicList
-            enableDragReorder
-            values={editWebsites}
-            placeholder="https://example.com"
-            onChange={setEditWebsites}
-            onChangeItem={(oldVal) => { if (oldVal) clearUrlWarning(oldVal); }}
-            onBlurItem={(val) => { if (val) updateUrlFormatWarning(val); }}
-            warnings={Object.fromEntries(
-              editWebsites
-                .map((v) => v.trim())
-                .filter(Boolean)
-                .flatMap((v) => {
-                  if (urlFormatWarnings[v]) return [[v, '⚠ Invalid URL']];
-                  if (urlDuplicates.has(v)) return [[v, '⚠ Already added']];
-                  return [];
-                })
-            )}
-          />
-          {user?.wayback_enabled !== 0 && user?.wayback_keys_configured !== 0 && (
-            <p className="form-field-hint">Wayback Machine archiving is enabled.</p>
-          )}
-        </div>
-
-        <div className="form-field">
-          <label className="form-label">Notes</label>
-          <textarea
-            className="form-textarea"
-            value={editNotes}
-            onChange={(e) => setEditNotes(e.target.value)}
-            rows={4}
-          />
-        </div>
-
-        <RssAlertPanel
-          editing
-          alertRssList={alertRssList}
-          newRssUrl={newRssUrl}
-          onNewRssUrlChange={setNewRssUrl}
-          onAddRss={handleAddRss}
-          onRemoveRss={handleRemoveRss}
-        />
-
-        <div className="detail-edit-actions-bottom">
-          <Button variant="primary" size="sm" onClick={handleSave} disabled={saving || !editName.trim() || Object.keys(emailFormatWarnings).length > 0 || Object.keys(phoneFormatWarnings).length > 0 || Object.keys(urlFormatWarnings).length > 0 || (!!newRssUrl.trim() && !isGoogleAlertUrl(newRssUrl.trim())) || (!!newRssUrl.trim() && alertRssList.some((f) => f.rss_url === newRssUrl.trim()))}>
-            {saving ? 'Saving…' : 'Save'}
-          </Button>
-          <Button variant="secondary" size="sm" onClick={() => setEditingAndNotify(false)} disabled={saving}>
-            Cancel
-          </Button>
-        </div>
-      </div>
+      <ContactEditForm
+        contact={contact}
+        user={user}
+        alertRssList={alertRssList}
+        newRssUrl={newRssUrl}
+        onNewRssUrlChange={setNewRssUrl}
+        onAddRss={handleAddRss}
+        onRemoveRss={handleRemoveRss}
+        onSaved={() => { onRefresh(); setEditingAndNotify(false); }}
+        onCancel={() => setEditingAndNotify(false)}
+      />
     );
   }
 
@@ -836,7 +358,7 @@ export default function GlobalTab({ contact, allProjects, onRefresh, onMembershi
               >
                 ↓ vCard
               </Button>
-              <Button variant="secondary" size="sm" onClick={startEdit}>Edit</Button>
+              <Button variant="secondary" size="sm" onClick={() => setEditingAndNotify(true)}>Edit</Button>
             </div>
             <Button variant="danger-outline" size="sm" onClick={() => setConfirmDelete(true)}>
               Delete contact
