+++ src/renderer/src/hooks/useSortFilter.ts
@@ -1,6 +1,8 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import type { SortDir } from '../views/ContactsTable';
 
+export type SortState<K extends string> = { key: K | null; dir: SortDir };
+
 export function useSortFilter<K extends string, F>(defaultFilters: F) {
   const [sort, setSort] = useState<{ key: K | null; dir: SortDir }>({ key: null, dir: 'asc' });
   const [filters, setFilters] = useState<F>(defaultFilters);
